perf: lazy load create telemetry (#5007)
